+++ src/main/java/Cup.java
@@ -1,3 +1,5 @@
+package main.java;
+
 import java.util.ArrayList;
 
 public class Cup {
