+++ src/main/java/Die.java
@@ -1,3 +1,5 @@
+package main.java;
+
 import java.util.Random;
 
 public class Die {
