+++ src/main/java/Main.java
@@ -1,3 +1,5 @@
+package main.java;
+
 public class Main {
 
 }
