+++ src/test/java/CupTest.java
@@ -1,3 +1,6 @@
+package test.java;
+
+import main.java.Cup;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
