+++ src/test/java/DieTest.java
@@ -1,3 +1,6 @@
+package test.java;
+
+import main.java.Die;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
